Libcorr

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: nchahed <nchahed@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/09 13:17:59 by nchahed           #+#    #+#             */
-/*   Updated: 2019/10/09 14:58:30 by nchahed          ###   ########.fr       */
+/*   Updated: 2019/10/09 16:24:07 by nchahed          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 	size_t	srclen;
 	size_t	dstlen;
